+++ src/trashing/trashinfo.rs
